Publish one hook for each verb
