+++ packages/ra-core/src/auth/index.ts
@@ -5,13 +5,26 @@ import useAuthState from './useAuthState';
 import usePermissions from './usePermissions';
 import useAuthenticated from './useAuthenticated';
 import WithPermissions from './WithPermissions';
+import useLogin from './useLogin';
+import useLogout from './useLogout';
+import useCheckAuth from './useCheckAuth';
+import useGetPermissions from './useGetPermissions';
 export * from './types';
+
 export {
     AuthContext,
+    useAuthProvider,
+    // low-vevel hooks for calling a particular verb on the authProvider
+    useLogin,
+    useLogout,
+    useCheckAuth,
+    useGetPermissions,
+    // hooks with state management
+    usePermissions,
+    useAuthState,
+    // hook with immediate effect
+    useAuthenticated,
+    // components
     Authenticated,
     WithPermissions,
-    useAuthProvider,
-    useAuthState,
-    useAuthenticated,
-    usePermissions,
 };
